+++ Sources/CoreCommands/SwiftTool.swift
@@ -50,13 +50,16 @@ import var TSCUtility.verbosity
 typealias Diagnostic = Basics.Diagnostic
 
 public struct ToolWorkspaceConfiguration {
+    let shouldInstallSignalHandlers: Bool
     let wantsMultipleTestProducts: Bool
     let wantsREPLProduct: Bool
 
     public init(
+        shouldInstallSignalHandlers: Bool = true,
         wantsMultipleTestProducts: Bool = false,
         wantsREPLProduct: Bool = false
     ) {
+        self.shouldInstallSignalHandlers = shouldInstallSignalHandlers
         self.wantsMultipleTestProducts = wantsMultipleTestProducts
         self.wantsREPLProduct = wantsREPLProduct
     }
@@ -308,57 +311,9 @@ public final class SwiftTool {
                 try ProcessEnv.chdir(packagePath)
             }
 
-            #if os(Windows)
-            // set shutdown handler to terminate sub-processes, etc
-            SwiftTool.cancellator = cancellator
-            _ = SetConsoleCtrlHandler({ _ in
-                // Terminate all processes on receiving an interrupt signal.
-                try? SwiftTool.cancellator?.cancel(deadline: .now() + .seconds(30))
-
-                // Reset the handler.
-                _ = SetConsoleCtrlHandler(nil, false)
-
-                // Exit as if by signal()
-                TerminateProcess(GetCurrentProcess(), 3)
-
-                return true
-            }, true)
-            #else
-            // trap SIGINT to terminate sub-processes, etc
-            signal(SIGINT, SIG_IGN)
-            let interruptSignalSource = DispatchSource.makeSignalSource(signal: SIGINT)
-            interruptSignalSource.setEventHandler {
-                // cancel the trap?
-                interruptSignalSource.cancel()
-
-                // Terminate all processes on receiving an interrupt signal.
-                try? cancellator.cancel(deadline: .now() + .seconds(30))
-
-                #if os(macOS) || os(OpenBSD)
-                // Install the default signal handler.
-                var action = sigaction()
-                action.__sigaction_u.__sa_handler = SIG_DFL
-                sigaction(SIGINT, &action, nil)
-                kill(getpid(), SIGINT)
-                #elseif os(Android)
-                // Install the default signal handler.
-                var action = sigaction()
-                action.sa_handler = SIG_DFL
-                sigaction(SIGINT, &action, nil)
-                kill(getpid(), SIGINT)
-                #else
-                var action = sigaction()
-                action.__sigaction_handler = unsafeBitCast(
-                    SIG_DFL,
-                    to: sigaction.__Unnamed_union___sigaction_handler.self
-                )
-                sigaction(SIGINT, &action, nil)
-                kill(getpid(), SIGINT)
-                #endif
+            if toolWorkspaceConfiguration.shouldInstallSignalHandlers {
+                cancellator.installSignalHandlers()
             }
-            interruptSignalSource.resume()
-            #endif
-
             self.cancellator = cancellator
         } catch {
             self.observabilityScope.emit(error)
